requirements.txt
asgiref==3.3.1
authentication==1.1.0
click==7.1.2
Django==3.1.7
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
passlib==1.7.4
pycryptodome==3.9.9
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
Werkzeug==1.0.1
WTForms==2.3.3

djangorestframework~=3.12.2